Tweak `unpackRepo()`, allow specifying an explicit `Path` for the resource
The `unpackRepo()` method is very useful for setting up example repo fixtures for
testing - it unzips zipped git archives on the resource classpath, and returns us
the JGit repository. However, loading resources from the classpath (while convenient,
meaning we don't need to worry about the absolute location on the filesystem of the
zip file) can also be temperamental, with the resource not being found if it's not
within the same classloader as the class of `com.madgag.git.test`.

It seems the most reliable way to be able to find our test resources is to specify
_which_ class we want to search the classpath of when we search for the resource. From
there, we can get an unambiguous path for the resource that can be used by
`unpackRepo()`.